Add support for returning raw json, and don't require an auth token
The public stash tab API returns very complicated results which are hard to model with a schema. Returning the json, and not type checking it, can be easier.

This also makes the auth token optional, and only sends it if it's set. GGG allows some access to their API without an auth token.