Add support for returning raw json, and don't require an auth token (#3)
* Add support for returning raw json, and don't require an auth token

The public stash tab API returns very complicated results which are hard to model with a schema. Returning the json, and not type checking it, can be easier. We may make the schema work properly, as it does type checking, but it's not necessary for a first version.

This also makes the auth token optional, and only sends it if it's set. GGG allows some access to their API without an auth token.

This also requires python3.8 and up, due to testing support. We can probably make it work with earlier python versions if necessary.